rag_project/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
langchain==0.3.14
langchain-community==0.3.14
langchain-groq==0.2.3
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.4
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
pypdf==5.1.0
